laravel class complete

+++ resources/views/photo/details.blade.php
@@ -1,19 +1,23 @@
 @extends('layouts/main')
 @section('content')
 
-<div class="callout primary">
-<article class="grid-container">
-<div class="">
+
+
+
+
+<div class="galleryBody">
 	<a href="/gallery/show/{{$photos->gallery_id}}">Back to portfolio</a>
 <h1>{{$photos->title}}</h1>
 <p class="lead">{{$photos->description}}</p>
 <p class="lead">{{$photos->location}}</p>
-</div>
-</article>
-</div>
 
-<div class="maindiv">
+
 <img class="proimage" src="/images/{{$photos->image}}" alt="">
-<p><a style="text-align:center" href="/photo/destroy/{{$photos->id}}/{{$photos->gallery_id}}">Delete Portfolio</a></p>
+<?php if (Auth::check()): ?>
+<p><a style="text-align:center" href="/photo/edit/{{$photos->id}}" >Edit Portfolio</a></p>
+<p><a style="text-align:center" href="/photo/destroy/{{$photos->id}}/{{$photos->gallery_id}}" onclick="return confirm('Are you sure?')">Delete Portfolio</a></p>
+
+<?php endif; ?>
+
 </div>
 @stop
